Rename Wambo\Catalog\Error to Wambo\Catalog\Exception
